+++ src/admin/views/Course/ListCourseView.tsx
@@ -5,6 +5,8 @@ import {
   DataTable,
   type DataTableAction,
   type DataTableColumn,
+  itemVariants,
+  containerVariants,
 } from "@/shared";
 import { useCourseView, type CourseResponseDto } from "@/admin";
 import styles from "./ListCourseView.module.css";
@@ -76,21 +78,6 @@ const ViewCoursesView: React.FC = () => {
       title: "Eliminar curso",
     },
   ];
-
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        staggerChildren: 0.1,
-      },
-    },
-  };
-
-  const itemVariants = {
-    hidden: { y: 20, opacity: 0 },
-    visible: { y: 0, opacity: 1 },
-  };
 
   return (
     <motion.div
