+++ src/admin/views/Year/ListYearView.tsx
@@ -6,6 +6,8 @@ import {
   DataTable,
   type DataTableAction,
   type DataTableColumn,
+  itemVariants,
+  containerVariants,
 } from "@/shared";
 import styles from "./ListYearView.module.css";
 import { useListYear } from "@/admin";
@@ -66,21 +68,6 @@ const ListYearView: React.FC = () => {
       title: "Eliminar año",
     },
   ];
-
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        staggerChildren: 0.1,
-      },
-    },
-  };
-
-  const itemVariants = {
-    hidden: { y: 20, opacity: 0 },
-    visible: { y: 0, opacity: 1 },
-  };
 
   return (
     <motion.div
