+++ src/shared/index.ts
@@ -2,3 +2,7 @@ export * from "./types";
 export * from "./utils";
 export * from "./hooks";
 export * from "./components";
+export {
+  itemVariants,
+  containerVariants,
+} from "./constants/variants.constants";
