+++ src/student/context/currentStudent/CurrentStudentContext.type.ts
@@ -1,4 +1,4 @@
-import type { Wallet } from "../../../admin/services/student/StudentService";
+import type { Wallet } from "@/admin";
 import type {
   CurrentStudent,
   AvatarComponents,
