+++ src/student/context/currentStudent/CurrentStudentProvider.tsx
@@ -10,7 +10,7 @@ import type {
 } from "../../types/CurrentStudent.type";
 import { useAuth } from "../../../user/hooks/useAuth";
 import { useHandleApiError } from "../../../shared/hooks/useHandleApiError";
-import type { Wallet } from "../../../admin/services/student/StudentService";
+import type { Wallet } from "@/admin";
 
 interface CurrentStudentProviderProps {
   children: ReactNode;
